% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_availability.R
\name{get_availability}
\alias{get_availability}
\title{Obtain availability in the UNCTAD TRAINS dataset.}
\usage{
get_availability()
}
\value{
\code{get_availability()} returns a data frame with nine character columns:
\itemize{
\item iso3_code: reporters three-letter countries codes;
\item country_code: reporters three-digit countries codes;
\item country_name: reporters names;
\item year;
\item nomen_code: reported nomenclature;
\item n_preferential_agreement: number of preferential agreement available;
\item partner_list: set of partners three-digit countries codes;
\item specific_duty: the value is 1 if the estimated tariff rate is
available for specific duty expression and 0 otherwise;
\item last_update: last updated date.
}
}
\description{
This function obtains the data availability of entire TRAINS dataset.

No parameters are required.
}
