% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_countries.R
\name{get_countries}
\alias{get_countries}
\title{Obtain information on countries that are available in the UNCTAD TRAINS
dataset.}
\usage{
get_countries()
}
\value{
\code{get_countries()} returns a data frame with six character columns:
\itemize{
\item iso3_code: three-letter countries codes;
\item country_code: three-digit countries codes;
\item name: countries' names;
\item is_reporter: the value is 1 if the country is available as reporter
and 0 otherwise;
\item is_partner: the value is 1 if the country is available as partner
and 0 otherwise;
\item is_group: the value is 1 if the code represents a group of countries
and 0 otherwise;
}
}
\description{
This function obtains all reporters and partners names, alpha numeric codes,
ISO3 codes and other meta data about the countries that are available in
the UNCTAD TRAINS dataset.

No parameters are required.
}
