% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nomen_codes.R
\name{get_nomen_codes}
\alias{get_nomen_codes}
\title{Obtain information on nomenclatures that are available in the UNCTAD TRAINS
dataset.}
\usage{
get_nomen_codes()
}
\value{
\code{get_nomen_codes()} returns a data frame with two character columns:
\itemize{
\item nomen_code: nomenclature code;
\item description: nomenclature description.
}
}
\description{
This function extracts information on nomenclatures that are available in the
UNCTAD TRAINS dataset.

No parameters are required.
}
