% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_products.R
\name{get_products}
\alias{get_products}
\title{Obtain information on products that are available in the UNCTAD TRAINS
dataset.}
\usage{
get_products()
}
\value{
\code{get_products()} returns a data frame with two character columns:
\itemize{
\item product_code: six-digit product codes (HS6);
\item product_desc: products' descriptions.
}
}
\description{
This function obtains the list of products available in the UNCTAD TRAINS
dataset.

No parameters are required.
}
