% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witstrainsr-package.R
\docType{package}
\name{witstrainsr}
\alias{witstrainsr}
\alias{witstrainsr-package}
\title{witstrainsr}
\description{
WITS REST API in R to programmatically access tariff data in UNCTAD TRAINS
database.
}
\details{
The complete and official WITS-API USER GUIDE can be accessed on
\url{http://wits.worldbank.org/data/public/WITSAPI_UserGuide.pdf}.
}
\author{
Diego Afonso de Castro \email{diegoacastro@gmail.com}
}
\keyword{internal}
